% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{constructGeneRange}
\alias{constructGeneRange}
\title{Find the start and end coordinates of the whole gene form joint exons.}
\usage{
constructGeneRange(joint_exon_ranges, flanking_length)
}
\description{
Find the start and end coordinates of the whole gene form joint exons.
}

