% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggleplotr.R
\name{generateTxStructurePlotData}
\alias{generateTxStructurePlotData}
\title{Plot read coverage across genomic regions}
\usage{
generateTxStructurePlotData(
  exons,
  cdss = NULL,
  transcript_annotations = NULL,
  rescale_introns = TRUE,
  new_intron_length = 50,
  flanking_length = c(50, 50),
  region_coords = NULL
)
}
\arguments{
\item{exons}{list of GRanges objects, each object containing exons for one transcript. 
The list must have names that correspond to transcript_id column in transript_annotations data.frame.}

\item{cdss}{list of GRanges objects, each object containing the coding regions (CDS) of a single transcript. 
The list must have names that correspond to transcript_id column in transript_annotations data.frame. 
If cdss is not specified then exons list will be used for both arguments. (default: NULL).}

\item{transcript_annotations}{Data frame with at least three columns: transcript_id, gene_name, strand. 
Used to construct transcript labels. (default: NULL)}

\item{rescale_introns}{Specifies if the introns should be scaled to fixed length or not. (default: TRUE)}

\item{new_intron_length}{length (bp) of introns after scaling. (default: 50)}

\item{flanking_length}{Lengths of the flanking regions upstream and downstream of the gene. (default: c(50,50))}

\item{region_coords}{Start and end coordinates of the region to plot, overrides flanking_length parameter.}
}
\value{
list of two objects 1) dataFrame 'transcript_struct_df' 2) vector 'limits' with length of 2
}
\description{
Also supports rescaling introns to constant length. Does not work 
on Windows, because rtracklayer cannot read BigWig files on Windows.
}
\examples{
require("dplyr")
require("GenomicRanges")

selected_transcripts = c("ENST00000438495", "ENST00000392477") #Plot only two transcripts of the gens
\dontrun{
tx_plot_data = generateTxStructurePlotData(ncoa7_exons[selected_transcripts], ncoa7_cdss[selected_transcripts])
# then the returned data can be plotted like
plotTranscriptStructure(exons_df = tx_plot_data$transcript_struct_df, limits = tx_plot_data$limits)
}

}
