% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getGenotypePalette}
\alias{getGenotypePalette}
\title{Returns a three-colour colour palette suitable for plotting coverage stratified by genotype}
\usage{
getGenotypePalette()
}
\description{
Returns a three-colour colour palette suitable for plotting coverage stratified by genotype
}

