% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.R
\name{makeCoveragePlot}
\alias{makeCoveragePlot}
\title{Make a Coverage plot of the region}
\usage{
makeCoveragePlot(
  coverage_df,
  limits,
  alpha = 1,
  fill_palette = c("#a1dab4", "#41b6c4", "#225ea8"),
  coverage_type = "area",
  show_genotype_legend = FALSE
)
}
\arguments{
\item{coverage_df}{the dataframe. Result of generateCoveragePlotData function}

\item{limits}{The cevtor of limits. Result of generateCoveragePlotData function}

\item{alpha}{Transparency (alpha) value for the read coverage tracks. 
Useful to set to something < 1 when overlaying multiple tracks (see track_id). (default: 1)}

\item{fill_palette}{Vector of fill colours used for the coverage tracks. Length must be equal to the number of 
unique values in track_data$colour_group column.}

\item{coverage_type}{Specifies if the read coverage is represented by either 'line', 'area' or 'both'. 
The 'both' option tends to give better results for wide regions. (default: area).}

\item{show_genotype_legend}{If TRUE, shows the legend of genotypes on the right of the plot. (default: false)}
}
\value{
gglot2 object
}
\description{
The coverage plot is compatible with wiggpleplotr Manhattan and transcript strucutre plots. 
Can be appended to those using the cowplot::plot_grid() function.
}
\examples{
coverage_plot_data = wiggleplotr::generateCoveragePlotData(...)
  coverage_plot = wiggleplotr::makeCoveragePlot(coverage_df = coverage_plot_data$coverage_df, 
                                                limits = coverage_plot_data$limits, 
                                                alpha = 1, 
                                                fill_palette = wiggleplotr::getGenotypePalette(), 
                                                coverage_type = "line", 
                                                show_genotype_legend = TRUE)
}
