% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{meanCoverage}
\alias{meanCoverage}
\title{Calculate mean coverage within each track_id and colour_group}
\usage{
meanCoverage(coverage_df)
}
\description{
Calculate mean coverage within each track_id and colour_group
}

