% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncoa7_cdss}
\alias{ncoa7_cdss}
\title{Coding sequences from 9 protein coding transcripts of NCOA7}
\format{A GRangesList object with 9 elements:
\describe{
  \item{element}{CDS start and end coordinates for a single transcript (GRanges object)}
  ...
}}
\source{
\url{http://www.ensembl.org/}
}
\usage{
ncoa7_cdss
}
\description{
A dataset containing start and end coordinates of coding sequences (CDS) from nine 
protein coding transcripts of NCOA7.
}
\keyword{datasets}

