% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncoa7_exons}
\alias{ncoa7_exons}
\title{Exons from 9 protein coding transcripts of NCOA7}
\format{
A GRangesList object with 9 elements:
\describe{
  \item{element}{Exon start and end coordinates for a single transcript (GRanges object)}
  ...
}
}
\source{
\url{http://www.ensembl.org/}
}
\usage{
ncoa7_exons
}
\description{
A dataset containing start and end coordinates of exons from nine 
protein coding transcripts of NCOA7.
}
\keyword{datasets}
