% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pasteFactors}
\alias{pasteFactors}
\title{Paste two factors together and preserved their joint order.}
\usage{
pasteFactors(factor1, factor2)
}
\arguments{
\item{factor1}{First factor}

\item{factor2}{Second factor}
}
\value{
Factors factor1 and factor2 pasted together.
}
\description{
Paste two factors together and preserved their joint order.
}
