% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.R
\name{plotTranscriptStructure}
\alias{plotTranscriptStructure}
\title{Make a Transcript structure plot}
\usage{
plotTranscriptStructure(
  exons_df,
  limits = NA,
  connect_exons = TRUE,
  xlabel = "Distance from gene start (bp)",
  transcript_label = TRUE
)
}
\arguments{
\item{exons_df}{the dataframe. Result of generateTxStructurePlotData function}

\item{limits}{The cevtor of limits. Result of generateTxStructurePlotData function (default: NA)}

\item{connect_exons}{Print lines that connect exons together. Set to FALSE when plotting peaks (default: TRUE).}

\item{xlabel}{Label to be shown in X-axis (default: "Distance from gene start (bp)").}

\item{transcript_label}{If TRUE then transcript labels are printed above each transcript. (default: TRUE).}
}
\value{
gglot2 object
}
\description{
The transcript structure plots is compatible with wiggpleplotr read coverage and manhattan plots. 
Can be appended to those using the cowplot::plot_grid() function.
}
\examples{
exon_plot_data <- wiggleplotr::generateTxStructurePlotData(..)
exon_plot <- wiggleplotr::plotTranscriptStructure(exons_df = exon_plot_data$transcript_struct_df, limits = exon_plot_data$limits)

}
