% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{subsamplePoints}
\alias{subsamplePoints}
\title{Choose a subsample of points to make plotting faster}
\usage{
subsamplePoints(tx_annotations, plot_fraction)
}
\description{
Makes sure that intron-exon boundaries are well samples.
}

