% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggleplotr-package.r
\docType{package}
\name{wiggleplotr}
\alias{wiggleplotr}
\alias{wiggleplotr-package}
\title{wiggleplotr}
\description{
wiggleplotr package provides tools to visualise transcript annotations (\code{\link[wiggleplotr]{plotTranscripts}}) and plot 
sequencing read coverage over annotated transcripts (\code{\link[wiggleplotr]{plotCoverage}}).
}
\details{
You can also use covenient wrapper functions 
(\code{\link[wiggleplotr]{plotTranscriptsFromEnsembldb}}), (\code{\link[wiggleplotr]{plotCoverageFromEnsembldb}}), 
(\code{\link[wiggleplotr]{plotTranscriptsFromUCSC}}) and (\code{\link[wiggleplotr]{plotCoverageFromUCSC}}).

To learn more about wiggleplotr, start with the vignette: 
\code{browseVignettes(package = "wiggleplotr")}
}
