% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getGenotypePalette}
\alias{getGenotypePalette}
\title{Returns a three-colour palette suitable for visualising read coverage stratified by genotype}
\usage{
getGenotypePalette(old = FALSE)
}
\arguments{
\item{old}{Return old colour palette (now deprecated).}
}
\value{
Vector of three colours.
}
\description{
Returns a three-colour palette suitable for visualising read coverage stratified by genotype
}
\examples{
getGenotypePalette()
}
