% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggleplotr.R
\name{plotTranscripts}
\alias{plotTranscripts}
\title{Function to quickly plot transcript structure without any read coverage}
\usage{
plotTranscripts(exons, cdss, annotations, rescale_introns = TRUE,
  new_intron_length = 50, flanking_length = c(50, 50))
}
\arguments{
\item{exons}{}

\item{cdss}{}

\item{annotations}{}

\item{rescale_introns}{}

\item{new_intron_length}{}

\item{flanking_length}{}
}
\value{
ggplot2 object
}
\description{
Function to quickly plot transcript structure without any read coverage
}

