% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggleplotr-package.r, R/wiggleplotr.R
\docType{package}
\name{wiggleplotr}
\alias{wiggleplotr}
\alias{wiggleplotr-package}
\title{wiggleplotr.}
\usage{
wiggleplotr(exons, cdss, track_data, transcript_annotations,
  rescale_introns = TRUE, new_intron_length = 50, flanking_length = c(50,
  50), plot_fraction = 0.1, heights = c(0.75, 0.25), alpha = 1,
  fill_palette = c("#a1dab4", "#41b6c4", "#225ea8"), mean_only = TRUE)
}
\arguments{
\item{exons}{List of GRanges objects, each object containing exons for one transcript. 
The list must have names that correspond to transcript_id column in transript_annotations data.frame.}

\item{cdss}{list of GRanges objects, each object containing the coding regions (CDS) of a single transcript. 
The list must have names that correspond to transcript_id column in transript_annotations data.frame. 
If specified then coding and non-coding regions will be plotted in a different colour.}

\item{track_data}{data.frame with the metadata for the bigWig read coverage files. Must contain the following columns:
\itemize{
 \item sample_id - unique id for each sample.
 \item track_id - if multiple samples (bigWig files) have the same track_id they will be overlayed on the same 
plot, track_id is also used as the facet label on the right.
 \item bigWig - path to the bigWig file.
 \item scaling_factor - normalisation factor for each sample, useful if different samples sequenced to different 
depth and bigWig files not normalised for that.
 \item colour_group - additional column to group samples into, is used as the colour of the coverage track.
}}

\item{transcript_annotations}{Data.frame with four columns: transcript_id, gene_id, gene_name, strand.}

\item{rescale_introns}{Specifies if the introns should be scaled to fixed length or not. (default: TRUE)}

\item{new_intron_length}{length (bp) of introns after scaling. (default: 50)}

\item{flanking_length}{Lengths of the flanking regions upstream and downstream of the gene. (default: c(50,50))}

\item{plot_fraction}{Size of the random sub-sample of points used to plot coverage (between 0 and 1). 
Smaller values make plotting significantly faster. (default: 0.1)}

\item{heights}{Specifies the proportion of the height that is dedicated to coverage plots (first value) 
relative to transcript annotations (second value). (default: c(0.75,0.25))}

\item{alpha}{Transparency (alpha) value for the read coverage tracks. 
Useful to set to something < 1 when overlaying multiple tracks (see track_id). (default: 1)}

\item{fill_palette}{Colour palette used for the coverage tracks.}

\item{mean_only}{Plot only mean coverage within each combination of track_id and colour_group values. 
Useful for example for plotting mean coverage stratified by genotype (which is specified in the colour_group column) (default: TRUE).}
}
\value{
Object form gridExtra::arrangeGrob() function. Can we viewed using plot command or saved using ggsave.
}
\description{
wiggleplotr.

Also supports rescaling introns to constant length.
}

